:truck: move CertifiedLevel to certif evaluation
