+++ api/src/certification/evaluation/domain/services/scoring/scoring-v2.js
@@ -17,13 +17,13 @@ import {
   AnswerCollectionForScoring,
   CertificationAssessmentScore,
   CertificationContract,
-  CertifiedLevel,
   CertifiedScore,
   CompetenceMark,
   ReproducibilityRate,
 } from '../../../../../shared/domain/models/index.js';
 import { AssessmentResultFactory } from '../../../../scoring/domain/models/factories/AssessmentResultFactory.js';
 import { AlgorithmEngineVersion } from '../../../../shared/domain/models/AlgorithmEngineVersion.js';
+import { CertifiedLevel } from '../../models/CertifiedLevel.js';
 
 /**
  * @param {Object} params
