+++ api/src/shared/domain/constants.js
@@ -109,7 +109,6 @@ const MINIMUM_REPRODUCIBILITY_RATE_TO_BE_CERTIFIED = 50;
 const MINIMUM_REPRODUCIBILITY_RATE_TO_BE_TRUSTED = 80;
 
 const V3_REPRODUCIBILITY_RATE = 100;
-const UNCERTIFIED_LEVEL = -1;
 
 const MAX_LEVEL_TO_BE_AN_EASY_TUBE = 3;
 const DEFAULT_LEVEL_FOR_FIRST_CHALLENGE = 2;
@@ -190,7 +189,6 @@ const constants = {
   PIX_COUNT_BY_LEVEL,
   PIX_ORIGIN,
   STUDENT_RECONCILIATION_ERRORS,
-  UNCERTIFIED_LEVEL,
 };
 
 export {
@@ -225,7 +223,6 @@ export {
   PIX_ORIGIN,
   STUDENT_RECONCILIATION_ERRORS,
   SUPPORTED_LOCALES,
-  UNCERTIFIED_LEVEL,
   V3_REPRODUCIBILITY_RATE,
   VALIDATION_ERRORS,
 };
