+++ api/tests/certification/evaluation/unit/domain/models/CertifiedLevel_test.js
@@ -1,9 +1,7 @@
-import {
-  MINIMUM_REPRODUCIBILITY_RATE_TO_BE_TRUSTED,
-  UNCERTIFIED_LEVEL,
-} from '../../../../src/shared/domain/constants.js';
-import { CertifiedLevel } from '../../../../src/shared/domain/models/CertifiedLevel.js';
-import { expect } from '../../../test-helper.js';
+import { CertifiedLevel } from '../../../../../../src/certification/evaluation/domain/models/CertifiedLevel.js';
+import { UNCERTIFIED_LEVEL } from '../../../../../../src/certification/shared/domain/constants.js';
+import { MINIMUM_REPRODUCIBILITY_RATE_TO_BE_TRUSTED } from '../../../../../../src/shared/domain/constants.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | CertifiedLevel', function () {
   context('when 3 challenges were answered', function () {
